requirements.txt
ultralytics
mysql-connector-python
python-dotenv